feat(board): add background, slug, and visibility settings forms
