+++ apps/web/src/views/board/index.tsx
@@ -54,7 +54,6 @@ import List from "./components/List";
 import { NewCardForm } from "./components/NewCardForm";
 import { NewListForm } from "./components/NewListForm";
 import { NewTemplateForm } from "./components/NewTemplateForm";
-import VisibilityButton from "./components/VisibilityButton";
 
 type PublicListId = string;
 
@@ -574,14 +573,6 @@ export default function BoardPage({ isTemplate }: { isTemplate?: boolean }) {
             )}
             {!isTemplate && (
               <>
-                <VisibilityButton
-                  visibility={boardData?.visibility ?? "private"}
-                  boardPublicId={boardId ?? ""}
-                  boardSlug={boardData?.slug ?? ""}
-                  queryParams={queryParams}
-                  isLoading={!boardData}
-                  isAdmin={workspace.role === "admin"}
-                />
                 {boardData && (
                   <Filters
                     labels={boardData.labels}
